feat: added sound file handling
The sound editor now has a per-sound section to select a sound file. If set, it takes precedence over the speech. The selected files are embedded into generated phonebooks as data URIs. When re-loading a file with embedded data URIs, the file name is not preserved and a label is shown next to the file input, indicating that there is an embedded file. Selecting a new file overrides the embedded one.

closes #2