fix: load and create phonebooks without any state properties
Fix a bug where generated phonebooks contained properties for the `any` state, and ignore such properties on opened phonebooks. Solution: when loading a phonebook, ignore any properties object defined on the state with ID `any`. When generating phonebooks, do not emit properties for the any state. The internal state now does not contain any such object, but the new getter with name `states` emits an any state with only a name and a description.

closes #7