feat: add run button to play phonebook remotely
Extend the toolbar with a select for an endpoint and a run button. The
select includes an "Add fernspielapparat..." option that opens up a
modal that can be used to establish a new connection.
Add a run button next to the select that can be used to establish a
connection and deploy the currently open phonebook, so that it runs
remotely. The active state on the selected remote is kept track of and
visually highlighted in the editor.