feat: add phonebook titles
Extend the phonebook format with a metadata section. Example section:
```
info:
  title: The amazing story of a cunning giraffe doctor
  author: John
  description: ""
```
The three properties describe the obvious things about this phonebook,
and do not change the semantics of the story when running on a fernspielapparat.
In a startup phonebook, both `title` and `author` get pseudo-random values.
`description` is currently always empty.
The title is shown in the toolbar section at the top and can be edited at will.
Downloaded phonebooks now have a generated filename derived from the title.