feat: add iteration number
Add an iteration number field to the phonebook format, which works a bit like a build hash. It is a counter that is incremented every time the phonebook can be observed to change by another application, that is, when the phonebook is either saved or deployed. The iteration number is stored in the phonebook info block, but is also used to suggest a filename.