fix: proper undo support for transitions
Transitions now support undo/redo. To make this easier, I did a massive refactoring to the transitions module. This involved making it TypeScript-based and adding a few tests. The internal format used by the module is now decoupled from the phonebook format. The internal format is now optimized for editing, and can be serialized to the phonebook format. Also changed build config so typescript works for lints, tests, etc.