test: better test coverage for transitions, fix some TS errors
The transition tests were out of sync with the new interfaces, so updated them to fit again, which fixes the TypeScript errors.
Added new tests for the transition getters, and removed testing of the omission of removed states during serialization, which is now left to getters.
Also using `expect.arrayContaining` now for cleaner matching on arrays.